requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
h5py==2.7.0
html5lib==0.9999999
Keras==2.0.6
Markdown==2.6.8
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.3.0
PyYAML==3.12
scipy==0.19.1
six==1.10.0
tensorflow==1.2.1
Theano==0.9.0
Werkzeug==0.12.2
